requirements.txt
python-telegram-bot==10.1.0
Pyrebase==3.0.27
feedparser==5.2.1
environs==2.1.1
